Add end of file check to loop
There is an inner loop that processes elements of a face line. This
commit replaces the unconditional while(true) loop with an end of file
check. It fixes a potential infinite loop condition when the input file
ends with a face line that is not terminated with a newline character.